Allow dataProvider to be a property or class constant.
getDataFromDataProviderAnnotation() will now first check to see whether
the named provider matches a method.  If it does, that method is used,
otherwise it will then check for a matching property (which may be
static or non-static), using that if it exists, and finally it will
check for a matching class constant.

If none of the above are present, an Exception is thrown.